requirements.txt
Click==7.0
Flask==1.0.3
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
PyYAML==5.1.1
Werkzeug==0.15.4
